# Conflicts:
#	CHANGELOG.md
#	base/AddFighterDlg.h
#	base/FightCategoryManager.cpp
#	base/FighterManager.cpp
#	base/FighterManager.h
#	base/FighterManagerDlg.cpp
#	base/MainWindowBase.cpp
#	base/base.pro
#	base/copy_files.cmd
#	base/ipponboard.qrc
#	build.cmd
#	doc/build_doc.cmd
#	scripts/create-versioninfo.cmd
#	setup/setup.iss
#	test/test.pro
#	util/SimpleCsvFile.hpp
#	util/path_helpers.h

+++ base/FighterManager.cpp
@@ -102,7 +102,7 @@ bool FighterManager::AddFighters(QString const& csvFile, QString& errorMsg)
 	for (unsigned int i=1; i<data.size(); ++i)
 	{
 		auto line = data[i];
-		if (line.count() != Specifiers.size())
+		if (line.count() != static_cast<int>(Specifiers.size()))
 		{
 			errorMsg = QObject::tr("Line %1 of file %2 has %3 fields but should have %4.").arg(i).arg(csvFile).arg(line.count()).arg(Specifiers.size());
 			return false;
@@ -144,7 +144,7 @@ bool FighterManager::SaveFighters(QString const& csvFile, QString& errorMsg)
 		return false;
 	}
 
-    qDebug("Successfully saved %i fighters to %s.", m_fighters.size(), csvFile.toStdString().c_str());
+    qDebug("Successfully saved %i fighters to %s.", static_cast<int>(m_fighters.size()), csvFile.toStdString().c_str());
 	return true;
 }
